requirements.txt
duckduckgo-search
fastapi
uvicorn
groq
phidata
yfinance
python-dotenv
openai
sqlalchemy
psycopg[binary]
pgvector
pypdf